+++ app/javascript/spec/dashboards/widget-wrapper.test.js
